refactor(test_sca): split pre-commit unit testing
